backend/requirements.txt
flask==2.3.2
flask-jwt-extended==4.4.4
python-dotenv==1.0.1
flask-cors==4.0.0
flask-limiter==3.5.0
openai==1.12.0
